requirements.txt
pandas
requests
sqlalchemy
snowflake-connector-python
snowflake-connector-python[pandas]
snowflake-sqlalchemy